Treat a vanished run folder as nobody's, and mirror the new test paths
A clear-up that finds the run folder already gone now answers "not mine"
instead of recreating a lock file in a folder someone else removed. Only
a genuine disk failure still throws.

Move the lock, record, and clean-up tests to the paths the mutation runner
mirrors from their sources, so a change to any of them selects its own
tests. Un-export the run-id prefix and the status schema, which had no
callers outside their module.
